Updated the Game Manager and fixed position bug
Made changes to the game manager so that it is the only thing we need to
place on each scene. It now creates the clock the first time it is
created and the close persists accross all scenes. Clock also hides when
in a dungeon.

Fixed a bug where the player would be posioned on top of the town
collider when returning to the wild.